pyproject.toml
[tool.poetry]
name = "house-price-simulator"
version = "0.1.0"
description = ""
authors = ["Diego Bravo <dbravo27@gmail.com>"]
readme = "README.md"
packages = [{include = "house_price_simulator"}]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
google-auth = "^2.16.1"
google-api-python-client = "^2.79.0"
requests = "^2.28.2"
djangorestframework = "^3.14.0"
django-crispy-forms = "^2.0"
django-environ = "^0.9.0"
django-debug-toolbar = "^3.8.1"
whitenoise = "^6.4.0"
numpy = "^1.24.2"
pandas = "^1.5.3"
matplotlib = "^3.7.0"
scrapy = "^2.8.0"
python-dotenv = "^1.0.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
sqlalchemy = "^2.0.7"
unidecode = "^1.3.6"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pylint = "^2.16.2"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
